Mark selections with a custom colour
Use a crazy hack to allow defining custom colour(s) for the selections
by using a "template" colour scheme file and copying into a subdir
using the name of the active colour scheme so it gets merged in to the
active scheme.

TODO: If we ever want to ship this, we'd need to elevate the
customisation to "user level" so they don't have to go digging around/
editing files in the package manually.